requirements.txt
altair
numpy
pandas
pydeck
streamlit
riskfolio-lib